refactor: чистка кода и обновление схемы
