+++ src/components/ExcelTable/Models/OptionEntity.ts
@@ -1,16 +1,21 @@
 import { DropdownOption } from 'components/ExcelTable/types';
+import { GeoObjectCategories } from 'generated/graphql';
 
 export class OptionEntity implements DropdownOption {
-  id: string;
+  private readonly _id: GeoObjectCategories;
 
   text: string;
 
   value: string;
 
-  constructor(id: string, text: string) {
-    this.id = id;
+  constructor(id: GeoObjectCategories, text: string) {
+    this._id = id;
     this.value = id;
     this.text = text;
+  }
+
+  get id(): GeoObjectCategories {
+    return this._id;
   }
 
   toString(): string {
