+++ src/utils/omitByTypename.ts
@@ -1,0 +1,3 @@
+import { omitAll } from 'lodash/fp';
+
+export const omitTypename = omitAll('__typename');
