+++ src/utils/packagingTableData/assembleAttributes.ts
@@ -1,0 +1,28 @@
+import { GridColumn } from 'components/ExcelTable/types';
+import { AttributeInput, ProjectStructureInput } from 'generated/graphql';
+import { defaultTo } from 'lodash/fp';
+import { omitTypename } from 'utils/omitByTypename';
+
+const removeCommas = (str: string) => str.split(',').filter(Boolean).join(',');
+
+export default function assembleAttributes(
+  columns: GridColumn[],
+  tableTemplate: ProjectStructureInput,
+): Array<AttributeInput> {
+  return columns.map(({ key, name }) => {
+    const attribute = tableTemplate.attributes.find(({ code }) => code === key);
+    const nameWithoutCommas = removeCommas(name);
+
+    return omitTypename(
+      defaultTo(
+        {
+          code: key,
+          name: nameWithoutCommas,
+          shortName: nameWithoutCommas,
+          units: '',
+        },
+        attribute,
+      ),
+    ) as AttributeInput;
+  });
+}
