+++ src/utils/packagingTableData/index.ts
@@ -1,0 +1,11 @@
+import assembleAttributes from './assembleAttributes';
+import assembleDomainEntities from './assembleDomainEntities';
+import assembleDomainObjects from './assembleDomainObjects';
+import assembleRisks from './assembleRisks';
+
+export {
+  assembleAttributes,
+  assembleDomainEntities,
+  assembleDomainObjects,
+  assembleRisks,
+};
